major clean up ./src ./build Makefile
