+++ src/dotenv.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+const char* get_env(const char* filename)
+{
+    FILE* fp = fopen(filename, 'r');
+    
+}
