+++ src/ilac.hxx
@@ -1,0 +1,15 @@
+#include "voucher_detail.hxx"
+#include "voucher.hxx"
+#include "pgutil.hxx"
+
+static std::string pad10 = "         ";
+static std::string pad5 = "     ";
+class console
+{
+    private:
+
+    public:
+        void display_voucher(int);
+
+       // auto spacer = []() { return "          ";};
+};
