+++ src/pg_finance_dao.h
@@ -1,0 +1,10 @@
+#ifndef PG_FINANCE_DAO_H
+#define PG_FINANCE_DAO_H
+
+#include "libpq-fe.h"
+
+
+PGconn PGsearch_vendor(char*);
+
+
+#endif
